# Faster `braille` printing