--- a/src/braille.jl
+++ b/src/braille.jl
@@ -1,8 +1,3 @@
-const BRAILLE_SYMBOLS = Char.(10240:10495)
-const BRAILLE_CODE = [1, 2, 4, 64, 8, 16, 32, 128]
-const H_BRAILLE = 4
-const W_BRAILLE = 2
-
 """
     braille(img; kwargs...)
     braille(img, alg; kwargs...)
@@ -35,7 +30,7 @@ function braille(img::GenericImage; kwargs...)
     return braille(img, DEFAULT_METHOD; kwargs...)
 end
 
-# Enable direct prining of Binary images:
+# Enable direct printing of Binary images:
 braille(img::AbstractMatrix{Bool}; kwargs...) = _braille(img; kwargs...)
 braille(img::BitMatrix; kwargs...) = _braille(img; kwargs...)
 function braille(img::AbstractMatrix{<:BinaryGray}; kwargs...)
@@ -44,41 +39,49 @@ end
 
 function _braille(A::AbstractMatrix; invert::Bool=false, to_string::Bool=false)
     invert && (A .= .!A)
-    B = _braille_matrix(A)
-    str = _mat2string(B)
-    to_string && return str
-    print(str)
+    to_string && return braille_string(A)
+    print_braille(A)
     return nothing
 end
 
-# Construct Matrix{Char} of Unicode Braille characters
-function _braille_matrix(A::AbstractMatrix)
-    hi, wi = size(A)
-    ho, wo = cld(hi, H_BRAILLE), cld(wi, W_BRAILLE)
-    out = Matrix{Char}(undef, ho, wo)
-
-    AP = PaddedView(0, A, (ho * H_BRAILLE, wo * W_BRAILLE))
-    for (i, tileaxs) in enumerate(TileIterator(axes(AP), (H_BRAILLE, W_BRAILLE)))
-        v = view(AP, tileaxs...)
-        out[i] = BRAILLE_SYMBOLS[_braille_index(v)]
-    end
-    return out
+function braille_string(A)
+    buff = IOBuffer()
+    print_braille(buff, A)
+    return String(take!(buff))
 end
 
-Base.@propagate_inbounds function _braille_index(A)
-    idx = 1
-    for i in 1:8
-        idx += A[i] * BRAILLE_CODE[i]
+hasrem(x, n) = !iszero(mod(x, n))
+roundup(x, n) = n * cld(x, n)
+
+print_braille(A) = print_braille(stdout, A)
+function print_braille(io::IO, A)
+    h, w = size(A)
+    if hasrem(h, 4) || hasrem(w, 2)
+        A = PaddedView(0, A, (roundup(h, 4), roundup(w, 2)))
     end
-    return idx
+    return _print_braille(io, A)
+end
+function _print_braille(io::IO, A)
+    rs, cs = axes(A)
+    for r in first(rs):4:last(rs)
+        for c in first(cs):2:last(cs)
+            @inbounds print(io, braille_symbol(view(A, r:(r + 3), c:(c + 1))))
+        end
+        println(io)
+    end
+    return io
 end
 
-function _mat2string(A::AbstractMatrix)
-    io = IOBuffer()
-    nmax = size(A, 1)
-    for (i, line) in enumerate(eachrow(A))
-        foreach(c -> print(io, c), line)
-        i < nmax && println(io)
-    end
-    return String(take!(io))
+function braille_symbol(A)
+    return @inbounds Char(
+        10240 +
+        A[1] +
+        2 * A[2] +
+        4 * A[3] +
+        64 * A[4] +
+        8 * A[5] +
+        16 * A[6] +
+        32 * A[7] +
+        128 * A[8],
+    )
 end