… test/references/braille/Bayer.txt
… test/references/braille/Bayer_invert.txt
… test/references/braille/FS.txt
… test/references/braille/FS_invert.txt
… test/references/gradient/Atkinson.txt
… test/references/gradient/BalancedCenteredPoint.txt
… test/references/gradient/Bayer.txt
… test/references/gradient/Bayer_invert_map.txt
… test/references/gradient/Bayer_l2.txt
… test/references/gradient/Bayer_l3.txt
… test/references/gradient/Bayer_l4.txt
… test/references/gradient/Bayer_linear.txt
… test/references/gradient/Burkes.txt
… test/references/gradient/CentralWhitePoint.txt
… test/references/gradient/ClosestColor.txt
… test/references/gradient/ClusteredDots.txt
… test/references/gradient/ConstantThreshold.txt
… test/references/gradient/FalseFloydSteinberg.txt
… test/references/gradient/Fan.txt
… test/references/gradient/FloydSteinberg.txt
… test/references/gradient/FloydSteinberg_clamp_error.txt
… test/references/gradient/IM_c5x5.txt
… test/references/gradient/IM_c6x6.txt
… test/references/gradient/IM_c7x7.txt
… test/references/gradient/IM_checks.txt
… test/references/gradient/IM_h4x4a.txt
… test/references/gradient/IM_h4x4o.txt
… test/references/gradient/IM_h6x6a.txt
… test/references/gradient/IM_h6x6o.txt
… test/references/gradient/IM_h8x8a.txt