#### Current release `v3.0.3`
```julia
julia> using DitherPunk

julia> A = rand(Bool, 1000, 1000);

julia> @benchmark braille($A; to_string=true)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  67.883 μs …  6.974 ms  ┊ GC (min … max): 0.00% … 98.66%
 Time  (median):     73.201 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   78.108 μs ± 94.868 μs  ┊ GC (mean ± σ):  1.69% ±  1.39%

  ▃▅▆██▇▆▅▃▃▂▁▁ ▁▂▁▁  ▁▁▃▁▂          ▁▁                       ▂
  █████████████████████████▆▇▇▆▆▆▇▇▇▇██▆██▇▆▇▆▅▆▅▄▁▁▅▆▅▅▅▄▁▃▄ █
  67.9 μs      Histogram: log(frequency) by time       136 μs <

 Memory estimate: 12.26 KiB, allocs estimate: 9.

julia> B = rand(Bool, 1000, 1000);

julia> @benchmark braille($B; to_string=true)
BenchmarkTools.Trial: 572 samples with 1 evaluation.
 Range (min … max):  6.939 ms … 41.132 ms  ┊ GC (min … max): 0.00% … 71.59%
 Time  (median):     7.974 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   8.700 ms ±  3.446 ms  ┊ GC (mean ± σ):  4.59% ±  9.10%

  ▃▇█▆▅▄▄▃▁                                                   
  ███████████▇▆▅▁▄▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄▁▁▄▁▁▁▁▄▁▁▁▁▁▄ ▇
  6.94 ms      Histogram: log(frequency) by time     27.7 ms <

 Memory estimate: 1.03 MiB, allocs estimate: 16.
```

#### This PR
```julia
julia> using DitherPunk

julia> A = rand(Bool, 100, 100);

julia> @benchmark braille($A; to_string=true)
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  46.578 μs …  4.979 ms  ┊ GC (min … max): 0.00% … 98.68%
 Time  (median):     48.045 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   51.144 μs ± 49.861 μs  ┊ GC (mean ± σ):  0.96% ±  0.99%

  ▂█▇▅▃▆▅▄▃▁▂                 ▂▁ ▃▃                      ▁▁   ▂
  █████████████▆█▇▇▆▆▆▅▄▆▄▅▆▄▇█████▇▆▇▆▅▇▇▇▆▆▅▄▄▆▅▃▄▅▇▄▁▄███▅ █
  46.6 μs      Histogram: log(frequency) by time      77.2 μs <

 Memory estimate: 7.20 KiB, allocs estimate: 8.

julia> B = rand(Bool, 1000, 1000);

julia> @benchmark braille($B; to_string=true)
BenchmarkTools.Trial: 907 samples with 1 evaluation.
 Range (min … max):  4.752 ms … 39.664 ms  ┊ GC (min … max): 0.00% … 84.57%
 Time  (median):     5.144 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.492 ms ±  1.664 ms  ┊ GC (mean ± σ):  2.29% ±  6.42%

       █▄                                                     
  ▄▆█▆████▅▅▄▄▃▃▄▄▃▃▃▃▃▃▃▃▃▂▃▃▃▃▂▂▂▂▂▂▂▂▂▂▁▂▂▁▂▂▂▂▂▂▂▂▂▁▂▁▁▂ ▃
  4.75 ms        Histogram: frequency by time        8.13 ms <

 Memory estimate: 570.70 KiB, allocs estimate: 14.

```